Ready4Hire/requirements.txt
# Ready4Hire v2.2 - Requirements
# Dependencias para el sistema completo

# ============================================================================
# Core Dependencies (Requeridas)
# ============================================================================

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.0.0
pydantic-settings>=2.0.0

# LLM & AI
ollama>=0.1.0
langchain>=0.0.350
langchain-community>=0.0.10

# GraphQL
strawberry-graphql[fastapi]>=0.216.0

# Embeddings & RAG
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4

# Database
asyncpg>=0.29.0
aioredis>=2.0.1

# PDF Generation
weasyprint>=60.0

# ML/DL (Core)
scikit-learn>=1.3.2
numpy>=1.24.0
torch>=2.1.0

# Utilidades
python-dotenv>=1.0.0
langid>=1.1.6

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
httpx>=0.25.0

# Security & Rate Limiting
slowapi>=0.1.9
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# ============================================================================
# Optional Dependencies (Para funcionalidades específicas)
# ============================================================================

# Fine-tuning (Solo si vas a entrenar modelos)
# Descomentar si tienes GPU con CUDA:
# unsloth @ git+https://github.com/unslothai/unsloth.git
# transformers>=4.35.0
# trl>=0.7.0
# peft>=0.6.0
# accelerate>=0.24.0
# bitsandbytes>=0.41.0

# Embeddings avanzados (Opcional)
# sentence-transformers>=2.2.2
# umap-learn>=0.5.5
# hdbscan>=0.8.33

# Audio processing (Opcional)
# openai-whisper>=20231117
# pyttsx3>=2.90

# Database (Opcional)
# asyncpg>=0.29.0
# aioredis>=2.0.1

# Code Quality (Desarrollo)
# black>=23.12.0
# flake8>=6.1.0
# mypy>=1.7.0
# pylint>=3.0.0
# bandit>=1.7.5

# ============================================================================
# Installation Instructions
# ============================================================================
#
# Quick install (básico):
#   pip install -r requirements.txt
#
# Con fine-tuning (requiere GPU):
#   pip install -r requirements.txt
#   pip install unsloth
#
# Completo (todas las features):
#   pip install -r requirements-full.txt
#
prometheus-client==0.19.0
